Linking fragment, waiting for section to be done to delete comments TODO
