+++ app/build.gradle
@@ -79,6 +79,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth'
     implementation 'com.google.mlkit:translate:17.0.0'
     implementation 'com.google.mlkit:language-id:17.0.3'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.13.2'
     testImplementation 'androidx.test.ext:junit:1.1.3'
     testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.0'
@@ -89,8 +90,8 @@ dependencies {
     androidTestImplementation 'androidx.test:rules:1.4.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     androidTestImplementation 'androidx.test.espresso:espresso-intents:3.4.0'
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.4.1'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.4.1'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.4.2'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.4.2'
     implementation platform('com.google.firebase:firebase-bom:29.1.0')
     implementation 'com.google.firebase:firebase-functions:20.0.2'
     implementation 'com.google.code.gson:gson:2.8.6'
