--- app/src/main/java/com/github/sdpsharelook/NavigationMenuActivity.kt
@@ -1,14 +1,0 @@
-package com.github.sdpsharelook
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-
-class NavigationMenuActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_navigation_menu)
-    }
-
-    fun back(view: View) = finish()
-}
