+++ app/src/main/java/com/github/sdpsharelook/language/LanguageSelectionDialog.kt
@@ -5,8 +5,8 @@ import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.widget.EditText
 import android.widget.ListView
-import androidx.appcompat.app.AppCompatActivity
 import androidx.core.widget.addTextChangedListener
+import androidx.fragment.app.FragmentActivity
 import com.github.sdpsharelook.R
 import com.github.sdpsharelook.translate.Translator
 import kotlin.coroutines.Continuation
@@ -14,7 +14,7 @@ import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
 class LanguageSelectionDialog private constructor(
-    private val activity: AppCompatActivity,
+    private val activity: FragmentActivity,
     private val languages: Set<Language>,
     private val translatorAvailable: Set<Language>?,
     private val ttsAvailable: Set<Language>?,
@@ -57,7 +57,7 @@ class LanguageSelectionDialog private constructor(
 
     companion object {
         suspend fun selectLanguage(
-            activity: AppCompatActivity,
+            activity: FragmentActivity,
             languages: Set<Language> = Translator.availableLanguages,
             translatorAvailable: Set<Language>? = null,
             ttsAvailable: Set<Language>? = null,
